Move core to workspace crate rune-core
It's key to note that in the process, we include different changes that
guide the build process of the crates, in order to arrive at a place
that is both ergonomic and easy to maintain:

We removed the `defvar`, `defvar_bool` and `defsym` macros. They were
not adding that much regarding work saved (just create a constant and
initialize it) and they were convoluting the dependency graph between
the crates. It's key that the core crate can use the symbols that were
previously generated on the build.rs file, but it does not need to know
the defuns.

For that, we moved the variables and symbols (not the defun symbols) to
a static file `env/sym.rs`. That file contains all the required
constants for the variables and symbols.

On the other hand, the defun symbols and constants are still generated
with the build script, to avoid loosing the convenience of defining the
functions with the `defun` proc macro.

Additionally, the macros were moved into their own module:
`rune_core::macros`, by re-exporting the now privately named macros. The
motivation behind it is that we need to access macros by name that may
conflict with other modules with the same name. An example of that is
the `error` module, which we import with self. The `error!` macro needs
to be imported as well, but a conflict arises. Now, we can import the
macro with `use rune_core::macros::error` and the module with `use rune_core::error`.